fix: Rewrite getPodBase to always get correct podBase
The new algorithm uses a combination of looking at the response headers to see if the current resource is the podBase, as defined by [`Clients can determine the storage of a resource by moving up the URI path hierarchy until the response includes a Link header with rel="type" targeting http://www.w3.org/ns/pim/space#Storage.`](https://solidproject.org/TR/protocol#storage), and at the same time also looking if there is a pim:Storage (as object) triple in the resource stating that this resource is the podBase; as some server implementations use this technique. Lastly, it also checks if there is a pim:storage (as predicate) triple in the resource, defining where the podBase is stored, used e.g. in the WebID resource on ESS.